requirements.txt
Babel==2.3.4
Beaker==1.8.0
beautifulsoup4==4.4.1
colander==1.2
debtcollector==1.3.0
futures==3.0.5
html2text==2016.1.8
iso8601==0.1.11
Jinja2==2.8
keystoneauth1==2.6.0
Mako==1.0.4
MarkupSafe==0.23
marrow.mailer==4.0.1
marrow.util==1.2.3
monotonic==1.0
msgpack-python==0.4.7
mysqlclient==1.3.7
netaddr==0.7.18
netifaces==0.10.4
nose==1.3.7
numpy==1.11.0
oslo.config==3.9.0
oslo.i18n==3.6.0
oslo.serialization==2.4.0
oslo.utils==3.8.0
pandas==0.18.0
passlib==1.6.5
PasteDeploy==1.5.2
pbr==1.9.1
positional==1.1.0
prettytable==0.7.2
Pygments==2.1.3
pyramid==1.6.1
pyramid-angularstarter==0.1
pyramid-beaker==0.8
pyramid-debugtoolbar==2.4.2
pyramid-jinja2==2.6.2
pyramid-mako==1.0.2
pyramid-marrowmailer==0.1
pyramid-tm==0.12.1
python-dateutil==2.5.2
python-keystoneclient==2.3.1
python-swiftclient==3.0.0
pytz==2016.3
repoze.lru==0.6
requests==2.9.1
six==1.10.0
SQLAlchemy==1.0.12
SQLAlchemy-Utils==0.32.0
stevedore==1.12.0
transaction==1.4.4
translationstring==1.3
venusian==1.0
waitress==0.8.10
WebError==0.13
WebOb==1.6.0
WebTest==2.0.20
wrapt==1.10.8
zope.deprecation==4.1.2
zope.interface==4.1.3
